# WIP, ENH: allow H5D_READS without opens