Very much a work in progress...

* allow the darshan runtime to increment `H5D_READS` without
assuming that the conventional `H5D` API was used to open
a file, as long as the return code from `H5Dread()` is non-negative.
(after all, how else could this return `0` if the file handle/dataset info
isn't properly available through some means that our runtime
cannot track?)

* this is based on a debugging/analysis of the `h5py` control
flow, which does not appear to call any of the conventional
`H5D` open/creation API functions (`H5F` is still accessed
through conventional means); instead, it appears that control
flow proceeds though the proxy function `dset_rw`, and darshan
then captures a call to `H5Dread()`

* considerably more work will be needed here I think, because
various other f(counters) are assumed to be valid but likely
are not in this case (like the timing for `H5D` events that
didn't happen);

* I also thought about trying to capture i.e., `dset_rw()`,
but it seems to me that there might be a sustainability issue
here--i.e., what if Rust and other languages were to have
"native" bindings that used yet another route, etc.? I guess
the question is whether we'd be happy to increment an open
event that is not using a conventional `H5D` API--my current intuition
is that this would not convey what is going on very accurately
(i.e., it is more accurate to say that darshan captured the read event
but didn't see the `H5D` open event through the conventional API)

* using this reproducer:

```python
from mpi4py import MPI
import numpy as np
from numpy.testing import assert_array_equal
import h5py

def main():
    comm = MPI.COMM_WORLD
    rank = comm.Get_rank()

    # round trip HDF5 IO test
    file_path = f"./test_rank_{rank}.hdf5"
    n_bytes = 10 * (rank + 1)

    bytes_to_write = np.ones(shape=n_bytes, dtype=np.int8)

    #with h5py.File(file_path, "w") as f:
        #f.create_dataset("dataset", data=bytes_to_write)

    print("before h5py.File()! on rank:", rank)
    with h5py.File(file_path, "r") as g:
        print("after h5py.File()! on rank:", rank)
        retrieved_data = np.asarray(g['dataset'])
        print("after retrieved_data on rank:", rank)
        assert_array_equal(retrieved_data, bytes_to_write)

if __name__ == "__main__":
    main()
```

* on this branch we now pick up the `H5D` read event with 3 ranks,
though the dataset ids or hashes darshan assigns need checking:

```
   rank                    id  H5F_OPENS  H5F_FLUSHES  H5F_USE_MPIIO
0     0   3991442081443699295          1            0              0
1     1  13842144362321976876          1            0              0
2     2  13503993907200906631          1            0              0
   rank                    id  H5D_OPENS  H5D_READS  H5D_WRITES  H5D_FLUSHES  H5D_BYTES_READ  ...  H5D_CHUNK_SIZE_D5  H5D_USE_MPIIO_COLLECTIVE  H5D_USE_DEPRECATED  H5D_FASTEST_RANK  H5D_FASTEST_RANK_BYTES  H5D_SLOWEST_RANK  H5D_SLOWEST_RANK_BYTES
0     0  17846303746652554285          0          1           0            0               0  ...                  0                         0                   0                 0                       0                 0                       0
1     1  13677892646774148885          0          1           0            0               0  ...                  0                         0                   0                 0                       0                 0                       0
2     2  17181974452690724755          0          1           0            0               0  ...                  0                         0                   0                 0                       0                 0                       0

[3 rows x 96 columns]
```

* on `main`:

```
   rank                    id  H5F_OPENS  H5F_FLUSHES  H5F_USE_MPIIO
0     0   3991442081443699295          1            0              0
1     1  13842144362321976876          1            0              0
2     2  13503993907200906631          1            0              0
Traceback (most recent call last):
  File "/home/tyler/LANL/rough_work/darshan/issue_690/check_for_h5dread.py", line 13, in <module>
    main(log_path=log_path)
  File "/home/tyler/LANL/rough_work/darshan/issue_690/check_for_h5dread.py", line 9, in main
    print(report.records["H5D"].to_df()["counters"])
KeyError: 'H5D'
```

* using the probe code:

```python
import sys

import darshan

def main(log_path: str):
    report = darshan.DarshanReport(log_path)
    print(report.records["H5F"].to_df()["counters"])
    print(report.records["H5D"].to_df()["counters"])

if __name__ == "__main__":
    log_path = sys.argv[1]
    main(log_path=log_path)
```