@tylerjereddy 2022-05-31
Note that I've also queried about the control flow in `h5py` upstream: https://github.com/h5py/h5py/issues/2107