@tylerjereddy 2022-06-05
We've now heard back from the folks upstream and I've added some detailed thoughts here: https://github.com/darshan-hpc/darshan/issues/690#issuecomment-1146837894