@tylerjereddy 2022-06-07
I'll close this and leave it to the team given the above-linked discussion--my main objective was to emphasize that I think we should have a more modular way to capture HDF5 symbols that is not reliant on specific API event chains.